setup ViewModel and use it with MainActivity
MainActivityViewModel was modified to include a dependency in form of
DataRepository class, which currently generates data itself, but in the
future will use other sources/choose where to get the data from so the
ViewModel never has to know how is the data obtained. The viewmodel holds
the reference to MutableLiveData obtained from the DataRepository and
provides a getter so the view can get the data in an immutable form, so
it doesn't modify it accidentally, which the view will only do via the
ViewModels methods.
MainActivity gets the ViewModelProvider.Factory injected, and obtains
the reference on it's viewmodel with the factory. It observes the data
obtained from the Viewmodel and changes it's views accordingly.